pyproject.toml
[project]
name = "whitespace"
version = "0.1.9"
description = "A self-hosted pastebin built with Flask. Supports syntax highlighting, file attachments, password-protected pastes, expiry options, and an admin panel."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.0.0",
    "flask-limiter>=3.5.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-wtf>=1.2.1",
    "gunicorn>=21.0.0",
    "pygments>=2.20.0",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
    "werkzeug>=3.0.0",
    "wtforms>=3.1.0",
]
